+++ CHANGELOG.md
@@ -1,4 +1,30 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2026-03-12 04:17:23
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 7fe744c86a4e3fbf86dd792cf2bad66bf9c44210
+
+## [ForwardWidgets-2kuai] - 2026-03-12 04:17:23
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 61fed6dc8ddd01551ec5891c4fe5354736d2cfa5
+
+## [Forward-xbzl] - 2026-03-12 04:17:23
+- 仓库: https://github.com/xbzl/Forward
+- 分支: main
+- 新增: dfd071a87b7f3b02d2aefec9fbae9ca072d32c33
+
+## [scripts-bemarkt] - 2026-03-12 04:17:23
+- 仓库: https://github.com/bemarkt/scripts
+- 分支: master
+- 新增: 535e2faac3f7aebe568aa6716503aa5ccf62b382
+
+## [ForwardWidgets-Madai-v] - 2026-03-12 04:17:23
+- 仓库: https://github.com/Madai-v/ForwardWidgets
+- 分支: main
+- 新增: 5827c84207950a95d93798f9de298f06864bb155
+
+
 ## [ForwardWidgets-huangxd] - 2026-03-11 04:13:05
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main